Calculate payouts

The function is a wrapper function for calc_payout_obs and calc_payout_dm, and exported to be used by the user. It calculates the expected payout to both observers and decision makers for each possible action undertaken by the decision maker, and the observers choice of whether or not to try and guess the rule.

calculate_payouts(
  entropy,
  pay_obs,
  pay_dm,
  pay_no_guess,
  deterministic = FALSE
)

Arguments

entropy
A list containing the entropy

pay_obs
A numeric with pay to the observer for guessing correctly

pay_dm
A numeric with pay to the decision maker if the observer does not guess

pay_no_guess
A numeric with pay to the observer for not guessing

deterministic
If TRUE a deterministic procedure is used to determine whether the observer tries to guess. Default is FALSE and the probability is calculated using a logit expression

Value

A list or list of lists where each list contains the payout to the observer and decision maker.